requirements.txt
allure_python_commons==2.13.2
cx_Oracle==8.3.0
cx_Oracle==8.3.0
Flask==2.3.2
gevent==22.10.1
h2==4.1.0
httpx==0.24.1
locustio==0.999
paramiko==3.1.0
prometheus_client==0.8.0
psycopg2==2.9.6
PyMySQL==0.10.1
pytest==7.4.0
PyYAML==6.0
PyYAML==6.0.1
six==1.12.0
spyne==2.14.0
xlrd==1.2.0
XlsxWriter==1.2.7
xlutils==2.0.0
xlwt==1.3.0
zeep==4.2.1
